Replace outdated custom Yii2 module with _before() method && update test suites configs
