+++ tests/_support/Helper/Functional.php
@@ -1,10 +1,15 @@
 <?php
+declare(strict_types = 1);
+
 namespace Helper;
 
 // here you can define custom actions
 // all public methods declared in helper class will be available in $I
 
-class Functional extends \Codeception\Module
-{
+use Codeception\Module;
 
+/**
+ * class Functional
+ */
+class Functional extends Module {
 }
